refactor(card-browser): move 'deleteSelectedNotes' to Fragment
